requirements.txt
access-points==0.2.45
Django==1.10.3
djangorestframework==3.5.3
MySQL-python==1.2.5
wireless==0.3.2
